feat: Update cell pixel dimensions to double for improved precision in rendering calculations.
De cel/pixel-berekening is nu echt op terminal-metrics gebaseerd in plaats van op een vaste of afgeronde verhouding. De normale ASCII-player vraagt nu de werkelijke character-cell pixels op via een aparte module: terminal_cell_metrics.cpp (line 160). ConsoleScreen gebruikt die waarde bij init en resize: consolescreen.cpp (line 1201).

Belangrijkste wijzigingen:

cellPixelWidth/cellPixelHeight zijn nu double door de ASCII sizing pipeline heen, zodat fractionele tekstvlakmetingen niet meteen kapot afgerond worden.
De oude vaste Radioify-windowfont-meting is uit de normale TUI-route gehaald.
Voor Windows Terminal gebruikt de code CSI 16t en daarna CSI 14t om echte pixelmaten te krijgen.
Voor klassieke consolehost vraagt de code de actuele Win32 consolefont-cell op.
De terminal-query zet alle niet-response inputrecords terug, zodat resize/key/mouse-events niet per ongeluk worden opgedronken.
De cover/aspect-berekening gebruikt nu de echte celpixels in playback_frame_output_ascii.cpp (line 208) en asciiart_layout.cpp (line 6).

Referenties: mpv’s aspect/viewport-aanpak in video/out/aspect.c, en Windows Terminal’s ondersteuning voor CSI 16t/CSI 14t pixelqueries in de Windows Terminal 1.22 release notes.

+++ playback/ascii/playback_frame_output.h
@@ -17,8 +17,8 @@ namespace playback_frame_output {
 // Keep ASCII/video-frame sizing deterministic; maxHeight is already clamped by caller.
 using AsciiOutputSizeCalculator =
     std::function<std::pair<int, int>(int width, int maxHeight, int srcW,
-                                      int srcH, int cellPixelWidth,
-                                      int cellPixelHeight)>;
+                                      int srcH, double cellPixelWidth,
+                                      double cellPixelHeight)>;
 using LogLineWriter = std::function<void(const std::string&)>;
 
 struct AsciiModePrepareInput {
@@ -29,8 +29,8 @@ struct AsciiModePrepareInput {
   bool allowAsciiCpuFallback = false;
   int width = 0;
   int maxHeight = 0;
-  int cellPixelWidth = 0;
-  int cellPixelHeight = 0;
+  double cellPixelWidth = 0.0;
+  double cellPixelHeight = 0.0;
   AsciiOutputSizeCalculator computeAsciiOutputSize;
   VideoFrame* frame = nullptr;
   AsciiArt* art = nullptr;
@@ -44,22 +44,22 @@ struct AsciiModePrepareInput {
   int* cachedMaxHeight = nullptr;
   int* cachedFrameWidth = nullptr;
   int* cachedFrameHeight = nullptr;
-  int* cachedCellPixelWidth = nullptr;
-  int* cachedCellPixelHeight = nullptr;
+  double* cachedCellPixelWidth = nullptr;
+  double* cachedCellPixelHeight = nullptr;
   LogLineWriter warningSink;
   LogLineWriter timingSink;
 };
 
 std::pair<int, int> computeAsciiPlaybackTargetSize(int width, int height,
                                                   int srcW, int srcH,
-                                                  int cellPixelWidth,
-                                                  int cellPixelHeight,
+                                                  double cellPixelWidth,
+                                                  double cellPixelHeight,
                                                   bool showStatusLine);
 
 std::pair<int, int> computeAsciiOutputSize(int maxWidth, int maxHeight,
                                           int srcW, int srcH,
-                                          int cellPixelWidth,
-                                          int cellPixelHeight);
+                                          double cellPixelWidth,
+                                          double cellPixelHeight);
 
 bool prepareAsciiModeFrame(AsciiModePrepareInput& input);
