+++ playback/ascii/playback_screen_renderer.h
@@ -51,8 +51,8 @@ struct PlaybackScreenRenderInputs {
   bool clearHistory = false;
   bool frameChanged = false;
   bool localSeekRequested = false;
-  int cellPixelWidth = 0;
-  int cellPixelHeight = 0;
+  double cellPixelWidth = 0.0;
+  double cellPixelHeight = 0.0;
   double pendingSeekTargetSec = -1.0;
   std::atomic<bool>* enableSubtitlesShared = nullptr;
   std::atomic<bool>* windowLocalSeekRequested = nullptr;
@@ -65,8 +65,8 @@ struct PlaybackScreenRenderInputs {
   int* cachedMaxHeight = nullptr;
   int* cachedFrameWidth = nullptr;
   int* cachedFrameHeight = nullptr;
-  int* cachedCellPixelWidth = nullptr;
-  int* cachedCellPixelHeight = nullptr;
+  double* cachedCellPixelWidth = nullptr;
+  double* cachedCellPixelHeight = nullptr;
   int* progressBarX = nullptr;
   int* progressBarY = nullptr;
   int* progressBarWidth = nullptr;
